reorganise definitions in headers
- Moved DecodingReturnValues to DecodingParams.hpp as HMM.hpp is not
actually a header file.
- Added ASMC.hpp for run() definition in main.cpp
